Huge remake of the code
I've only done it on the one with AI because its more interesting than the other one. 
What's new is the fact that it's not only a text-based game but now using Pygame it's 2d.
It's still not finished and the AI doesnt really work, but I wanted to just put it out now and update it to it's final (or maybe not) version when I finish it.